+++ RecipeBox.API/Models/Rating.cs
@@ -3,7 +3,7 @@ namespace RecipeBox.API.Models
     public class Rating
     {
         public int RatingId { get; set; }
-        public int Score { get; set; }
+        public double Score { get; set; }
         public User Rater { get; set; }
         public int RaterId { get; set; }
         public Post Post { get; set; }
